Added new methods to create and read data outside CRUD requests.

+++ Concessionario.cs
@@ -23,9 +23,9 @@
         public double Weight { get; set; }
         public double Speed { get; set; }
         public double MaxSpeed { get; set; }
-        public int Displacement { get; set; }
+        public double Displacement { get; set; }
 
-        public Car(int id, string plate, string name, string brand, double weight, double speed, double maxSpeed, int displacement)
+        public Car(int id, string plate, string name, string brand, double weight, double speed, double maxSpeed, double displacement)
         {
             Id = id;
             Plate = plate;
